add flavors for running cucumber and junit instrumentation tests
turns out that the cucumber test runner and the junit test runner are mutually exclusive.
this means, we can't run cucumber tests and "classic" instrumentations tests in the same run.
as a workaround, I added two product flavors. When running instrumentation tests, choose the right
flavor for the desired tests to be run. It was not possible to achieve the behavior with tasks.